move SMSVerify, added NewPassword layout, added error inputs in mock, fix capital letter i18n text
