+++ src/components/FormInputs/FormInputs.module.scss
@@ -2,5 +2,4 @@
     display: flex;
     flex-direction: column;
     gap: 20px;
-    padding: 0 21px;
 }
