+++ src/i18n.tsx
@@ -50,8 +50,8 @@ i18n
                             phoneNumber: 'Телефон',
                             placeholder: '+7 (***)',
                             continue: 'Далее',
-                            ValidationError: 'Введен некорректный номер телефона.',
-                            UserWithThatPhoneNotFound: 'Пользователь с данным номером телефона не существует.',
+                            validationError: 'Введен некорректный номер телефона.',
+                            userWithThatPhoneNotFound: 'Пользователь с данным номером телефона не существует.',
                         },
 
                         newPassword: {
@@ -59,7 +59,7 @@ i18n
                             nameLabel: 'Новый пароль',
                             nameLabelRepeat: 'Повторите пароль',
                             button: 'Сохранить',
-                            ValidationError: 'Некорректный пароль',
+                            validationError: 'Некорректный пароль',
                             passwordDontMatch: 'Пароли не совпадают',
                             invalidCredentials: 'Пароль введен неверно, повторите попытку еще раз.',
                         },
@@ -137,7 +137,7 @@ i18n
                             server: 'Сервер не отвечает. Попробуйте позже',
                             validation: 'Данные заполненые неверно.',
                             duplicate: 'Такой номер телефона уже зарегистрирован',
-                            ValidationError: 'Некорректный ввод',
+                            validationError: 'Некорректный ввод',
                             invalidConformationCode: 'Введен неправильный код',
                         },
                         restaurants: {
