+++ src/pages/RestorePassword/NewPassword/NewPassword.module.scss
@@ -1,0 +1,11 @@
+.new_password {
+    &__layout {
+        padding: 0 21px;
+        display: flex;
+        flex-direction: column;
+
+        :first-child {
+            margin-bottom: 20px;
+        }
+    }
+}
